AP-4263: Refactor CFECivil classes
Extract shared methods and remove superfluous methods from
new ObtainStateBenefitTypesService as a lot were
shared with other CFELegacy submission services
